fast hack for showing images (deleted items are not handled yet)

+++ app/src/main/kotlin/com/ydanneg/erply/ui/screens/main/catalog/Catalog.kt
@@ -21,7 +21,6 @@ import androidx.compose.material3.pulltorefresh.rememberPullToRefreshState
 import androidx.compose.material3.rememberDrawerState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -50,7 +49,7 @@ private fun ProductGroupPreview() {
             parentId = "0",
             order = index,
             description = LocalizedValue("description$item"),
-            changed = (System.currentTimeMillis() / 1000).toInt()
+            changed = System.currentTimeMillis() / 1000
         )
     }
